requirements.txt
numpy==1.21.6
pandas==1.4.0
scikit_learn==1.0.2
umap_learn==0.5.3
